+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Size prop to Modal.
 - `aria-label`, `aria-labelledby`, and `aria-describedby` prop to Modal.
 - `FocusTrap` in modal.
+- Focus on the previous Element when Modal is closed.
 
 ### Fixed